pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "portainer-tui"
version = "0.1.0"
description = "A terminal UI for managing Portainer instances"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "jn-frdn" }]
keywords = ["portainer", "docker", "tui", "terminal", "devops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Terminals",
]
requires-python = ">=3.11"
dependencies = [
    "textual>=0.89.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.1",
    "python-dotenv>=1.0",
]

[project.urls]
Homepage = "https://github.com/jn-frdn/portainer-tui"
Repository = "https://github.com/jn-frdn/portainer-tui"
Issues = "https://github.com/jn-frdn/portainer-tui/issues"
Changelog = "https://github.com/jn-frdn/portainer-tui/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.11",
    "textual-dev>=1.6",
    "types-pyyaml",
    "respx>=0.21",
]

[project.scripts]
portainer-tui = "portainer_tui.__main__:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["portainer_tui"]
